# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Meetup Football API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/matches": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get list of matches",
                "operationId": "c0735c7be1ec75cab53dd45bfed420ef",
                "responses": {
                    "200": {
                        "description": "List of matches"
                    }
                }
            },
            "post": {
                "tags": [
                    "Matches"
                ],
                "summary": "Create a new match",
                "operationId": "6d1c46be4db252cc42bd1a24914b76bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "starts_at",
                                    "duration",
                                    "match_type",
                                    "max_players",
                                    "location_name",
                                    "address",
                                    "city"
                                ],
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Partido de futbol casual"
                                    },
                                    "starts_at": {
                                        "type": "string",
                                        "example": "2026-04-01 18:00:00"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 90
                                    },
                                    "match_type": {
                                        "type": "string",
                                        "example": "7v7"
                                    },
                                    "max_players": {
                                        "type": "integer",
                                        "example": 14
                                    },
                                    "required_level": {
                                        "type": "string",
                                        "enum": [
                                            "beginner",
                                            "intermediate",
                                            "advanced"
                                        ]
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 5
                                    },
                                    "location_name": {
                                        "type": "string",
                                        "example": "Camp Municipal"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Carrer Gran Via 1"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Barcelona"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Match created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/matches/{footballMatch}": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get a match by ID",
                "operationId": "65ff32a940e20c21994a325890150c37",
                "parameters": [
                    {
                        "name": "footballMatch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match details"
                    },
                    "404": {
                        "description": "Match not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "Update a match",
                "operationId": "9c25d3b62e762ec2f29985ed0fb4bc5a",
                "parameters": [
                    {
                        "name": "footballMatch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "Partido modificado"
                                    },
                                    "starts_at": {
                                        "type": "string",
                                        "example": "2026-04-01 18:00:00"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 90
                                    },
                                    "match_type": {
                                        "type": "string",
                                        "example": "7v7"
                                    },
                                    "max_players": {
                                        "type": "integer",
                                        "example": 14
                                    },
                                    "required_level": {
                                        "type": "string",
                                        "enum": [
                                            "beginner",
                                            "intermediate",
                                            "advanced"
                                        ]
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 5
                                    },
                                    "location_name": {
                                        "type": "string",
                                        "example": "Camp Municipal"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Carrer Gran Via 1"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Barcelona"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "open",
                                            "full",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Match updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Match not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Matches"
                ],
                "summary": "Delete a match",
                "operationId": "14612e734d23d8cf6ad32497c4ae1d65",
                "parameters": [
                    {
                        "name": "footballMatch",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Match deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Match not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}